+++ src/lib/core/scope.ts
@@ -2,8 +2,9 @@ import {
     type IRenderable,
     type IZodTagRenderable,
     createRenderable,
-    isRenderable
+    isZTRenderable
 } from "./renderable";
+import { isSchemaType } from "./schema";
 
 /** Extracts kargs from scope for zod schemas */
 export const scopedKargs = (value: any, kargs: any) => (value?._zod && value.__ztScope) ? kargs[value.__ztScope] : kargs;
@@ -18,10 +19,10 @@ export const withScope = (vals: any[], scope?: string) => {
     if (!scope) return vals;
     return vals.map(v => {
         let _v = v as any;
-        if (isRenderable(v)) {
+        if (isZTRenderable(v)) {
             // create scoped renderable
             _v = scopedRenderable(v, scope);
-        } else if (v?._zod) {
+        } else if (isSchemaType(v)) {
             // create scoped schema
             _v = v.clone();
             Object.defineProperty(_v, '__ztScope', {